backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
RPi.GPIO==0.7.1
pytz==2024.1
